requirements.txt
Flask==1.0.2
hypothesis==3.78.0
pytest==5.1.2
pytest-cov==2.7.1
pytest-xdist==1.29.0
mock==3.0.5
flaky==3.6.1
pytest-html==2.0.0
requests-mock==1.7.0
flask-testing==0.7.1
locustio==0.11.0
jsonschema==2.6.0